// src/Code.enum.ts
export enum Code {
    // Alphanumeric section
    KeyA = "KeyA",
    KeyB = "KeyB",
    KeyC = "KeyC",
    KeyD = "KeyD",
    KeyE = "KeyE",
    KeyF = "KeyF",
    KeyG = "KeyG",
    KeyH = "KeyH",
    KeyI = "KeyI",
    KeyJ = "KeyJ",
    KeyK = "KeyK",
    KeyL = "KeyL",
    KeyM = "KeyM",
    KeyN = "KeyN",
    KeyO = "KeyO",
    KeyP = "KeyP",
    KeyQ = "KeyQ",
    KeyR = "KeyR",
    KeyS = "KeyS",
    KeyT = "KeyT",
    KeyU = "KeyU",
    KeyV = "KeyV",
    KeyW = "KeyW",
    KeyX = "KeyX",
    KeyY = "KeyY",
    KeyZ = "KeyZ",
    Digit0 = "Digit0",
    Digit1 = "Digit1",
    Digit2 = "Digit2",
    Digit3 = "Digit3",
    Digit4 = "Digit4",
    Digit5 = "Digit5",
    Digit6 = "Digit6",
    Digit7 = "Digit7",
    Digit8 = "Digit8",
    Digit9 = "Digit9",

    // Modifier keys
    AltLeft = "AltLeft",
    AltRight = "AltRight",
    CapsLock = "CapsLock",
    ControlLeft = "ControlLeft",
    ControlRight = "ControlRight",
    MetaLeft = "MetaLeft",
    MetaRight = "MetaRight",
    ShiftLeft = "ShiftLeft",
    ShiftRight = "ShiftRight",

    // Whitespace keys
    Enter = "Enter",
    Space = "Space",
    Tab = "Tab",

    // Navigation keys
    ArrowDown = "ArrowDown",
    ArrowLeft = "ArrowLeft",
    ArrowRight = "ArrowRight",
    ArrowUp = "ArrowUp",
    End = "End",
    Home = "Home",
    PageDown = "PageDown",
    PageUp = "PageUp",

    // Function keys
    Escape = "Escape",
    F1 = "F1",
    F2 = "F2",
    F3 = "F3",
    F4 = "F4",
    F5 = "F5",
    F6 = "F6",
    F7 = "F7",
    F8 = "F8",
    F9 = "F9",
    F10 = "F10",
    F11 = "F11",
    F12 = "F12",

    // Numeric keypad keys
    NumLock = "NumLock",
    Numpad0 = "Numpad0",
    Numpad1 = "Numpad1",
    Numpad2 = "Numpad2",
    Numpad3 = "Numpad3",
    Numpad4 = "Numpad4",
    Numpad5 = "Numpad5",
    Numpad6 = "Numpad6",
    Numpad7 = "Numpad7",
    Numpad8 = "Numpad8",
    Numpad9 = "Numpad9",
    NumpadAdd = "NumpadAdd",
    NumpadComma = "NumpadComma",
    NumpadDecimal = "NumpadDecimal",
    NumpadDivide = "NumpadDivide",
    NumpadEnter = "NumpadEnter",
    NumpadEqual = "NumpadEqual",
    NumpadMultiply = "NumpadMultiply",
    NumpadSubtract = "NumpadSubtract",

    // Symbols
    Backquote = "Backquote",
    Backslash = "Backslash",
    BracketLeft = "BracketLeft",
    BracketRight = "BracketRight",
    Comma = "Comma",
    Equal = "Equal",
    Minus = "Minus",
    Period = "Period",
    Quote = "Quote",
    Semicolon = "Semicolon",
    Slash = "Slash",

    // International keys
    IntlBackslash = "IntlBackslash",
    IntlRo = "IntlRo",
    IntlYen = "IntlYen",
    Lang1 = "Lang1",
    Lang2 = "Lang2",
    Lang3 = "Lang3",
    Lang4 = "Lang4",
    Lang5 = "Lang5",

    // Other special keys
    Delete = "Delete",
    Insert = "Insert",
    Pause = "Pause",
    PrintScreen = "PrintScreen",
    ScrollLock = "ScrollLock",

    // Media keys
    AudioVolumeDown = "AudioVolumeDown",
    AudioVolumeMute = "AudioVolumeMute",
    AudioVolumeUp = "AudioVolumeUp",
    MediaPlayPause = "MediaPlayPause",
    MediaSelect = "MediaSelect",
    MediaStop = "MediaStop",
    MediaTrackNext = "MediaTrackNext",
    MediaTrackPrevious = "MediaTrackPrevious",

    // Browser control keys
    BrowserBack = "BrowserBack",
    BrowserFavorites = "BrowserFavorites",
    BrowserForward = "BrowserForward",
    BrowserHome = "BrowserHome",
    BrowserRefresh = "BrowserRefresh",
    BrowserSearch = "BrowserSearch",
    BrowserStop = "BrowserStop",

    // Power control keys
    Power = "Power",
    Sleep = "Sleep",
    WakeUp = "WakeUp",

    // Additional F-keys and OEM-specific keys
    // Include all 661 keys based on MDN documentation

    Backspace = "Backspace",
    F13 = "F13",
    F14 = "F14",
    F15 = "F15",
    F16 = "F16",
    F17 = "F17",
    F18 = "F18",
    F19 = "F19",
    F20 = "F20",
    F21 = "F21",
    F22 = "F22",
    F23 = "F23",
    KanaMode = "KanaMode",
    F24 = "F24",
    Convert = "Convert",
    NonConvert = "NonConvert",
    LaunchApp2 = "LaunchApp2",
    LaunchApp1 = "LaunchApp1",
    LaunchMail = "LaunchMail",
    ContextMenu = "ContextMenu",
    Again = "Again",
    Undo = "Undo",
    Select = "Select",
    Copy = "Copy",
    Open = "Open",
    Paste = "Paste",
    Find = "Find",
    Cut = "Cut",
    Help = "Help",
    Eject = "Eject",
    Props = "Props",
    Fn = "Fn"
}